refactor(server): reorganize sync/ by transport into ws/ and http/ subdirs
Group sync files by their transport dependency cluster instead of flat.
Delete deprecated createSyncPlugin (near-identical copy of createWsSyncPlugin).
Migrate tests to createWsSyncPlugin, remove tests for deprecated REST doc routes.
