+++ packages/server/src/sync/index.ts
@@ -5,8 +5,7 @@
  * CRITICAL: This file must NOT import from `@epicenter/workspace` — it's the dependency firewall.
  */
 
-export { createHttpSyncPlugin, type HttpSyncPluginConfig } from './http-sync-plugin';
-export { createWsSyncPlugin, type WsSyncPluginConfig } from './ws-sync-plugin';
+export { createHttpSyncPlugin, type HttpSyncPluginConfig } from './http/plugin';
 export {
 	type SyncStorage,
 	compactDoc,
@@ -14,7 +13,5 @@ export {
 	decodeSyncRequest,
 	encodeSyncRequest,
 	stateVectorsEqual,
-} from './storage';
-
-/** @deprecated Use `createWsSyncPlugin` instead. */
-export { createSyncPlugin, type SyncPluginConfig } from './plugin';
+} from './http/storage';
+export { createWsSyncPlugin, type WsSyncPluginConfig } from './ws/plugin';
