+++ packages/server/src/sync/ws/rooms.test.ts
