+++ packages/server/src/sync/ws/rooms.ts
